Configuration for OpenMP

+++ test/utils/file/read_from_file_test.cc
@@ -49,7 +49,7 @@ void ReadLinesFromFile::TearDown() {
 TEST_F(ReadLinesFromFile, Expect_Result_ToContainTheContentOfTheFileWhenTheFileExists) {
   std::vector<std::string> result;
   result = read_from_file::ReadLinesFromFile(kMultipleLinesReadFileName);
-  for (int i = 0; i < result.size(); i ++)
+  for (size_t i = 0; i < result.size(); i ++)
     EXPECT_EQ(kMultipleLinesListExpected, result) << std::endl;
 }
